M    src/moduls/daq/JavaLikeCalc/freefunc.cpp
git-svn-id: svn://oscada.org/trunk/OpenSCADA@2032 7bc60720-51fb-0310-bb35-cf765b5d350c

+++ src/moduls/daq/JavaLikeCalc/freefunc.cpp
@@ -642,7 +642,7 @@ Reg *Func::cdTypeConv( Reg *op, Reg::Type tp, bool no_code )
 			//    *rez = (int)rez->val().r_el;
 			    break;
 			case Reg::String:
-			    *rez = (*rez->val().s_el != EVAL_STR) ? strtoll(rez->val().s_el->c_str(),NULL,0) : EVAL_INT;
+			    *rez = (*rez->val().s_el != EVAL_STR) ? (int64_t)strtoll(rez->val().s_el->c_str(),NULL,0) : EVAL_INT;
 			    break;
 			default: break;
 		    }
